add framework prompt for `grida code x` command
